requirements.txt
opencv-python
flask
google-search
wikipedia
argparse
google-cloud